# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Indian Language Localizer API",
    "version": "1.0.0",
    "description": "Comprehensive AI-powered multilingual translation and localization backend for 22 Indian languages using IndicTrans2, LLaMA 3, and Whisper Large V3",
    "contact": {
      "name": "API Support",
      "email": "support@indianlocalizer.ai"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.indianlocalizer.ai",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/translate": {
      "post": {
        "tags": ["Translation"],
        "summary": "Translate content",
        "description": "Translate text content with cultural localization using IndicTrans2 and LLaMA 3",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Translation completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchTranslationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters or unsupported language",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Translation engine failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/speech/stt": {
      "post": {
        "tags": ["Speech"],
        "summary": "Speech to Text (STT)",
        "description": "Convert speech audio to text using Whisper Large V3",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Audio file (WAV, MP3, MP4, M4A, OGG, FLAC)"
                  },
                  "language": {
                    "type": "string",
                    "description": "Optional language hint for better accuracy",
                    "example": "hi"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Speech-to-text conversion successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/STTResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid language or request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "File too large (max 100 MB)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported audio format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/speech/tts": {
      "post": {
        "tags": ["Speech"],
        "summary": "Text to Speech (TTS)",
        "description": "Convert text to speech audio using VITS/Tacotron2",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TTSRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Audio file generated successfully",
            "content": {
              "audio/mpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid language or text parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "TTS generation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/speech/localize": {
      "post": {
        "tags": ["Speech"],
        "summary": "Audio localization pipeline",
        "description": "Complete audio localization: STT → Translation → TTS",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Audio file to localize"
                  },
                  "target_language": {
                    "type": "string",
                    "description": "Target language code",
                    "example": "hi"
                  },
                  "domain": {
                    "type": "string",
                    "description": "Domain for localization context",
                    "example": "healthcare",
                    "default": "general"
                  }
                },
                "required": ["file", "target_language"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Audio localization completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioLocalizationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters or unsupported language",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TranslationRequest": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "Text content to translate",
            "example": "Hello, how are you?"
          },
          "file_id": {
            "type": "string",
            "description": "Alternative to text: translate content from uploaded file"
          },
          "source_language": {
            "type": "string",
            "description": "Source language code (en or Indian language)",
            "example": "en"
          },
          "target_languages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of target language codes",
            "example": ["hi", "ta", "bn"]
          },
          "domain": {
            "type": "string",
            "description": "Domain for cultural localization",
            "enum": ["general", "healthcare", "construction", "technical", "legal", "education"],
            "example": "general",
            "default": "general"
          },
          "apply_localization": {
            "type": "boolean",
            "description": "Apply cultural and domain-specific localization",
            "default": true
          }
        },
        "required": ["source_language", "target_languages"]
      },
      "BatchTranslationResponse": {
        "type": "object",
        "properties": {
          "translations": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string"
                },
                "confidence": {
                  "type": "number",
                  "format": "float"
                },
                "localized": {
                  "type": "boolean"
                }
              }
            },
            "example": {
              "hi": {
                "text": "नमस्ते, आप कैसे हैं?",
                "confidence": 0.92,
                "localized": true
              },
              "ta": {
                "text": "வணக்கம், நீங்கள் எப்படி இருக்கிறீர்கள்?",
                "confidence": 0.89,
                "localized": true
              }
            }
          },
          "source_text": {
            "type": "string",
            "example": "Hello, how are you?"
          },
          "source_language": {
            "type": "string",
            "example": "en"
          },
          "domain": {
            "type": "string",
            "example": "general"
          },
          "translation_time": {
            "type": "number",
            "format": "float",
            "example": 1.23
          },
          "model_used": {
            "type": "string",
            "example": "IndicTrans2"
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "STTResponse": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "नमस्ते, मेरा नाम राहुल है"
          },
          "language": {
            "type": "string",
            "example": "hi"
          },
          "confidence": {
            "type": "number",
            "format": "float",
            "example": 0.94
          },
          "duration": {
            "type": "number",
            "format": "float",
            "example": 3.2
          },
          "processing_time": {
            "type": "number",
            "format": "float",
            "example": 0.8
          },
          "model_used": {
            "type": "string",
            "example": "whisper-large-v3"
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "TTSRequest": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "Text to convert to speech",
            "example": "नमस्ते, आपका स्वागत है"
          },
          "language": {
            "type": "string",
            "description": "Language code for TTS",
            "example": "hi"
          },
          "voice": {
            "type": "string",
            "description": "Voice type",
            "enum": ["male", "female"],
            "default": "female"
          }
        },
        "required": ["text", "language"]
      },
      "AudioLocalizationResponse": {
        "type": "object",
        "properties": {
          "original_text": {
            "type": "string",
            "example": "Welcome to our healthcare center"
          },
          "translated_text": {
            "type": "string",
            "example": "हमारे स्वास्थ्य केंद्र में आपका स्वागत है"
          },
          "source_language": {
            "type": "string",
            "example": "en"
          },
          "target_language": {
            "type": "string",
            "example": "hi"
          },
          "audio_file": {
            "type": "string",
            "example": "/storage/outputs/audio_localized_hi_1234567890.mp3"
          },
          "processing_time": {
            "type": "number",
            "format": "float",
            "example": 4.5
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string",
            "example": "Error message description"
          },
          "error_code": {
            "type": "string",
            "example": "TRANSLATION_FAILED"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:00:00Z"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}